source/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlmodel==0.0.14
sqlalchemy==2.0.28
python-jose==3.3.0
passlib[bcrypt]==1.7.4
reportlab==4.0.8
pydantic==1.10.14
python-multipart==0.0.9
